# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend Go untuk manga library: scan/move folder, bookmark, translate, bug report, duplikat.\nLogin lewat POST /login, lalu klik Authorize dan isi `Bearer \u003caccess_token\u003e`.",
        "title": "Manga Library API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/duplicates": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "duplicates"
                ],
                "summary": "List kandidat duplikat yang belum direview",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Halaman",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Item per halaman",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/dto.DuplicateCandidateItem"
                                    }
                                },
                                "page": {
                                    "type": "integer"
                                },
                                "per_page": {
                                    "type": "integer"
                                },
                                "total_data": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/duplicates/backfill": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "duplicates"
                ],
                "summary": "Konversi tabrakan nama lama jadi kandidat duplikat",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Model.BaseResponseModel"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/dto.BackfillResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/duplicates/{id}/compare": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "duplicates"
                ],
                "summary": "Bandingkan halaman existing vs incoming",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "duplicate_candidates.id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Model.BaseResponseModel"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/dto.DuplicateCompareResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Model.BaseResponseModel"
                        }
                    }
                }
            }
        },
        "/duplicates/{id}/resolve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "action: new_title (butuh new_title), merge (butuh merge_mode replace|append), keep_existing.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "duplicates"
                ],
                "summary": "Selesaikan kandidat duplikat",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "duplicate_candidates.id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Aksi",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ResolveDuplicateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Model.BaseResponseModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Model.BaseResponseModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Model.BaseResponseModel": {
            "type": "object",
            "properties": {
                "Code": {
                    "type": "integer"
                },
                "Data": {},
                "HeaderMessage": {
                    "type": "string"
                },
                "Message": {
                    "type": "string"
                }
            }
        },
...
        "dto.BackfillResult": {
            "type": "object",
            "properties": {
                "converted": {
                    "type": "integer"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "dto.DuplicateCandidateItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "existing_new_folder_id": {
                    "type": "integer"
                },
                "existing_page_count": {
                    "type": "integer"
                },
                "existing_thumbnail": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "incoming_page_count": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.DuplicateComparePage": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "dto.DuplicateCompareResponse": {
            "type": "object",
            "properties": {
                "existing_pages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DuplicateComparePage"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "incoming_pages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DuplicateComparePage"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dto.ResolveDuplicateRequest": {
            "type": "object",
            "required": [
                "action"
            ],
            "properties": {
                "action": {
                    "type": "string"
                },
                "merge_mode": {
                    "type": "string"
                },
                "new_title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Format: `Bearer \u003caccess_token\u003e`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}